+++ ios/Plugin/PhotoViewer.swift
@@ -10,6 +10,12 @@ enum PhotoViewerError: Error {
     var oneImageViewController: OneImageViewController?
     var sliderViewController: SliderViewController?
     var stFrom: Int = 0
+    private var config: PhotoViewerConfig
+    
+    init(config: PhotoViewerConfig) {
+        self.config = config
+        super.init()
+    }
 
     // MARK: collectionController
 
@@ -65,5 +71,55 @@ enum PhotoViewerError: Error {
         } else {
             return false
         }
+    }
+    
+    @objc public func saveImageFromHttpToInternal(_ call: CAPPluginCall, url: String,
+                                                  fileName: String) throws {
+        // get image location from the config
+        guard let imageLocation = config.iosImageLocation else {
+            let message = "You must have 'iosImageLocation' defined in " +
+                          "the capacitor.config.ts file"
+            throw PhotoViewerError.failed(message: message)
+        }
+        guard let imageURL = URL(string: url) else {
+            let message = "could not convert image url to URL"
+            throw PhotoViewerError.failed(message: message)
+        }
+
+        UtilsImage.downloadAndSaveImage(imageURL: imageURL,
+                                       imageName: fileName,
+                             imageLocation: imageLocation) { result in
+            switch result {
+            case .success(let imagePath):
+                // convert the filepath into a Web View-friendly path.
+                if let range = imagePath.path.range(of: "/Containers/", options: .backwards) {
+                    let extractedPath = imagePath.path[range.upperBound...]
+                    let resultPath = "capacitor://localhost/_capacitor_file_" +
+                                     "/var/mobile/Containers/" + String(extractedPath)
+                    call.resolve(["webPath": resultPath])
+                } else {
+                    call.resolve(["message": "cannot create the Web View-friendly path"])
+                }
+            case .failure(let error):
+                call.reject("Error: \(error)")
+                return
+            }
+        }
+
+    }
+    @objc public func getInternalImagePaths() throws -> [String] {
+        // get image location from the config
+        guard let imageLocation = config.iosImageLocation else {
+            let message = "You must have 'iosImageLocation' defined in " +
+                          "the capacitor.config.ts file"
+            throw PhotoViewerError.failed(message: message)
+        }
+        do {
+            let pathList = try UtilsImage.listOfImagePath(for: imageLocation)
+            return pathList
+        } catch let error {
+            throw PhotoViewerError.failed(message: error.localizedDescription)
+        }
+        
     }
 }