+++ ios/Plugin/PhotoViewerConfig.swift
@@ -1,0 +1,11 @@
+//
+//  PhotoViewerConfig.swift
+//  Plugin
+//
+//  Created by  Quéau Jean Pierre on 22/09/2023.
+//  Copyright © 2023 Max Lynch. All rights reserved.
+//
+
+public struct PhotoViewerConfig {
+    var iosImageLocation: String?
+}